Refactor translations
